after review

+++ internal/server/handlers/delete.go
@@ -13,11 +13,18 @@ func (a Api) DeleteBook(w http.ResponseWriter, r *http.Request) {
 	idB, err := strconv.Atoi(mux.Vars(r)["id"])
 	if err != nil {
 		log.Println("error mux.Vars book in delete.go")
+		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
 
 	id := int64(idB)
 	ctx := r.Context()
+
+	if id == 0 {
+		w.WriteHeader(http.StatusBadRequest)
+		log.Println("error, not ID in request")
+		return
+	}
 
 	err = a.Storage.DeleteBook(ctx, id)
 	if err != nil {
